requirements.txt
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
langchain==0.0.350
numpy==1.26.2
openai==1.5.0
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.5.2